Get rid of the warning, but do it right.  Since the vector of unsigned ints really does get assigned -1 sometimes (and that on purpose), it should really be a vector of int.  That avoids the doubtful (unsigned int)-1.
[[Imported from SVN: r6656]]

+++ dune/grid/uggrid/uggridfactory.cc
@@ -184,7 +184,7 @@ void Dune::GridFactory<Dune::UGGrid<dimworld> >::
 insertBoundarySegment(const std::vector<unsigned int>& vertices,
                       const shared_ptr<BoundarySegment<dimworld> > boundarySegment)
 {
-  array<unsigned int, dimworld*2-2> segmentVertices;
+  array<int, dimworld*2-2> segmentVertices;
 
   for (size_t i=0; i<vertices.size(); i++)
     segmentVertices[i] = vertices[i];
@@ -329,7 +329,7 @@ createGrid()
 
       int numVertices = (dimworld==2)
                         ? 2
-                        : ((boundarySegmentVertices_[i][3]==(unsigned int)(-1)) ? 3 : 4);
+                        : ((boundarySegmentVertices_[i][3]==-1) ? 3 : 4);
 
       double segmentCoordinates[dimworld*2-2][dimworld];
       for (int j=0; j<numVertices; j++)
